*fix dds loading under linux *todo add dynamic display modes list under linux
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@1444 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/ref_gl/r_local.h
@@ -1252,30 +1252,30 @@ void R_FlipImage(int idx, img_t* pix, byte* dst);
 #define DDSCAPS2_CUBEMAP	0x00000200
 
 typedef struct {
-	unsigned long dwSize;
-	unsigned long dwFlags;
-	unsigned long dwFourCC;
-	unsigned long dwRGBBitCount;
-	unsigned long dwRBitMask;
-	unsigned long dwGBitMask;
-	unsigned long dwBBitMask;
-	unsigned long dwABitMask;
+	unsigned int dwSize;
+	unsigned int dwFlags;
+	unsigned int dwFourCC;
+	unsigned int dwRGBBitCount;
+	unsigned int dwRBitMask;
+	unsigned int dwGBitMask;
+	unsigned int dwBBitMask;
+	unsigned int dwABitMask;
 } ddsFilePixelFormat_t;
 
 typedef struct
 {
-	unsigned long dwSize;
-	unsigned long dwFlags;
-	unsigned long dwHeight;
-	unsigned long dwWidth;
-	unsigned long dwPitchOrLinearSize;
-	unsigned long dwDepth;
-	unsigned long dwMipMapCount;
-	unsigned long dwReserved1[11];
+	unsigned int dwSize;
+	unsigned int dwFlags;
+	unsigned int dwHeight;
+	unsigned int dwWidth;
+	unsigned int dwPitchOrLinearSize;
+	unsigned int dwDepth;
+	unsigned int dwMipMapCount;
+	unsigned int dwReserved1[11];
 	ddsFilePixelFormat_t ddspf;
-	unsigned long dwCaps1;
-	unsigned long dwCaps2;
-	unsigned long dwReserved2[3];
+	unsigned int dwCaps1;
+	unsigned int dwCaps2;
+	unsigned int dwReserved2[3];
 } ddsFileHeader_t;
